test: extract validation tests into individual file
