+++ packages/custom-field/test/custom-field.test.js
@@ -103,71 +103,6 @@ describe('custom field', () => {
     });
   });
 
-  describe('required', () => {
-    beforeEach(() => {
-      customField.required = true;
-    });
-
-    it('should set invalid to false by default', () => {
-      expect(customField.invalid).to.be.false;
-    });
-
-    it('should become invalid on validate call when empty', () => {
-      expect(customField.invalid).to.be.false;
-      customField.validate();
-      expect(customField.invalid).to.be.true;
-    });
-
-    it('should become valid after receiving a non-empty value from "change" event', () => {
-      customField.inputs[0].value = 'foo';
-      dispatchChange(customField.inputs[0]);
-      expect(customField.invalid).to.be.false;
-    });
-
-    it('should become invalid after receiving an empty value from "change" event', () => {
-      customField.value = 'foo';
-      customField.inputs[0].value = '';
-      dispatchChange(customField.inputs[0]);
-      expect(customField.invalid).to.be.true;
-    });
-  });
-
-  describe('validation', () => {
-    it('should check validity on validate', () => {
-      const spy = sinon.spy(customField, 'checkValidity');
-      customField.validate();
-      expect(spy.called).to.be.true;
-    });
-
-    it('should run validation on input change', () => {
-      const spy = sinon.spy(customField, 'checkValidity');
-      customField.inputs[0].value = 'foo';
-      dispatchChange(customField.inputs[0]);
-      expect(spy.called).to.be.true;
-    });
-
-    it('should fire a validated event on validation success', () => {
-      const validatedSpy = sinon.spy();
-      customField.addEventListener('validated', validatedSpy);
-      customField.validate();
-
-      expect(validatedSpy.calledOnce).to.be.true;
-      const event = validatedSpy.firstCall.args[0];
-      expect(event.detail.valid).to.be.true;
-    });
-
-    it('should fire a validated event on validation failure', () => {
-      const validatedSpy = sinon.spy();
-      customField.addEventListener('validated', validatedSpy);
-      customField.required = true;
-      customField.validate();
-
-      expect(validatedSpy.calledOnce).to.be.true;
-      const event = validatedSpy.firstCall.args[0];
-      expect(event.detail.valid).to.be.false;
-    });
-  });
-
   describe('custom parser and formatter', () => {
     it('should use custom parser if that exists', async () => {
       customField.parseValue = (value) => {
