+++ packages/email-field/test/validation-lit.test.js
@@ -1,2 +1,2 @@
 import '../src/vaadin-lit-email-field.js';
-import './email-field.common.js';
+import './validation.common.js';
