+++ packages/email-field/test/validation-polymer.test.js
@@ -1,2 +1,2 @@
 import '../src/vaadin-email-field.js';
-import './email-field.common.js';
+import './validation.common.js';
