+++ packages/password-field/test/password-field.test.js
@@ -208,87 +208,6 @@ describe('password-field', () => {
       expect(revealButton.hasAttribute('aria-hidden')).to.be.false;
     });
   });
-
-  describe('initial validation', () => {
-    let validateSpy;
-
-    beforeEach(() => {
-      passwordField = document.createElement('vaadin-password-field');
-      validateSpy = sinon.spy(passwordField, 'validate');
-    });
-
-    afterEach(() => {
-      passwordField.remove();
-    });
-
-    it('should not validate by default', async () => {
-      document.body.appendChild(passwordField);
-      await nextRender();
-      expect(validateSpy.called).to.be.false;
-    });
-
-    it('should not validate when the field has an initial value', async () => {
-      passwordField.value = 'Initial Value';
-      document.body.appendChild(passwordField);
-      await nextRender();
-      expect(validateSpy.called).to.be.false;
-    });
-
-    it('should not validate when the field has an initial value and invalid', async () => {
-      passwordField.value = 'Initial Value';
-      passwordField.invalid = true;
-      document.body.appendChild(passwordField);
-      await nextRender();
-      expect(validateSpy.called).to.be.false;
-    });
-  });
-
-  describe('validation', () => {
-    it('should fire a validated event on validation success', () => {
-      const validatedSpy = sinon.spy();
-      passwordField.addEventListener('validated', validatedSpy);
-      passwordField.validate();
-
-      expect(validatedSpy.calledOnce).to.be.true;
-      const event = validatedSpy.firstCall.args[0];
-      expect(event.detail.valid).to.be.true;
-    });
-
-    it('should fire a validated event on validation failure', () => {
-      const validatedSpy = sinon.spy();
-      passwordField.addEventListener('validated', validatedSpy);
-      passwordField.required = true;
-      passwordField.validate();
-
-      expect(validatedSpy.calledOnce).to.be.true;
-      const event = validatedSpy.firstCall.args[0];
-      expect(event.detail.valid).to.be.false;
-    });
-  });
-});
-
-describe('invalid', () => {
-  let field;
-
-  beforeEach(() => {
-    field = fixtureSync('<vaadin-password-field invalid></vaadin-password-field>');
-  });
-
-  it('should not remove "invalid" state when ready', () => {
-    expect(field.invalid).to.be.true;
-  });
-});
-
-describe('invalid with value', () => {
-  let field;
-
-  beforeEach(() => {
-    field = fixtureSync('<vaadin-password-field invalid value="123456"></vaadin-password-field>');
-  });
-
-  it('should not remove "invalid" state when ready', () => {
-    expect(field.invalid).to.be.true;
-  });
 });
 
 describe('disabled', () => {
